Remove globally applied transition
This avoids jankiness when lightbox is opened/closed as it changes all
the elements by applying the hidden attribute. So when the user closes
the lightbox a transition happens because the hidden attribute is then
removed. This causes the page to "shift" as the transition is applied
and everything gets laid out on screen.