Add Github Action for pushing to Github Pages
The Dockerfile in the root of the repository is used to build an image
which then gets used by the GitHub Actions to generate the static
website into a gh-pages branch which is deployed via GitHub Pages. Upon
pushing to the master branch the Hugo build will get triggered.

This action will eventually run the build commands within the docker
container and deploy the website to a GitHub Pages branch.